+++ src/typescript/macros/index.ts
@@ -1,6 +1,9 @@
 import { executeSkillRoll } from "./skillRoll.js";
 import { executeSpecialRoll } from "./specialRoll.js";
-import { executeWeaponAttack } from "./weaponAttack.js";
+import {
+  executeWeaponAttack,
+  executeWeaponAttackFromSource
+} from "./weaponAttack.js";
 
 /** An object to hold references to system Macro command functions. */
 export const macros = {
@@ -19,10 +22,19 @@ export const macros = {
   executeSpecialRoll,
 
   /**
-   * Execute a Weapon Attack macro on an Actor.
-   * @param actorId - the ID of the Actor, the weapon belongs to
-   * @param weaponId - the ID of the Weapon on the Actor
-   * @param attackName - the name of the Attack on the weapon
+   * Execute a Weapon Attack Macro.
+   * @param weaponId - the ID of the Weapon
+   * @param attackName - the name of the Attack on the Weapon
+   * @param actorId - the ID of the Actor, the Weapon belongs to, blank if
+   *                  unowned
    */
-  executeWeaponAttack
+  executeWeaponAttack,
+
+  /**
+   * Execute a Weapon Attack from the provided Weapon source.
+   * @param data - the source data of the weapon
+   * @param attackName - the name of the weapon attack in the source data
+   * @param options - additional roll options
+   */
+  executeWeaponAttackFromSource
 };