+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - attacks on unowned weapons will now first use the selected token, then the
   impersonated actor of the user to prepopulate their data
 - ability to create macros from unowned weapons
+- ability to create weapon macros from source provided in a script macro
+  ([#60](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/60))
 
 ### Changed